refactor: update user schema and task entity to use new imports and enhance relationships
