+++ drizzle/meta/_journal.json
@@ -8,6 +8,13 @@
       "when": 1750876809163,
       "tag": "0000_cute_old_lace",
       "breakpoints": true
+    },
+    {
+      "idx": 1,
+      "version": "7",
+      "when": 1751096827229,
+      "tag": "0003_amusing_raza",
+      "breakpoints": true
     }
   ]
 }
