+++ src/auth/get-user.decorator.ts
@@ -1,5 +1,5 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from './user.entity';
+import { createParamDecorator, ExecutionContext } from "@nestjs/common";
+import { User } from "./schema";
 
 export const GetUser = createParamDecorator(
   (_data, ctx: ExecutionContext): User => {
