+++ src/auth/jwt.strategy.ts
@@ -4,10 +4,9 @@ import { PassportStrategy } from "@nestjs/passport";
 import { InjectRepository } from "@nestjs/typeorm";
 import { ExtractJwt, Strategy } from "passport-jwt";
 import { JwtPayload } from "./jwt-payload.interface";
-import * as schema from "./schema";
+import {User} from "./schema";
 import { UsersRepository } from "./user.repository";
 
-type User = typeof schema.users.$inferSelect;
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
