+++ src/auth/schema.ts
@@ -1,4 +1,6 @@
+import { relations } from "drizzle-orm";
 import { pgTable, text, uuid, varchar } from "drizzle-orm/pg-core";
+import { reviews } from "@/reviews/schema";
 import { timestamps } from "@/utils/columns";
 
 export const users = pgTable("user", {
@@ -7,6 +9,11 @@ export const users = pgTable("user", {
   email: varchar("email", { length: 255 }).notNull().unique(),
   password: varchar("password", { length: 32 }).notNull(),
   picture: text("picture").notNull(),
-  refreshToken: text("refresh_token").nullable(),
   ...timestamps,
 });
+
+export const reviewRelations = relations(reviews, ({ one }) => ({
+  user: one(users, { fields: [reviews.user_id], references: [users.id] }),
+}));
+
+export type User = typeof users.$inferSelect;
