+++ src/tasks/task.entity.ts
@@ -1,11 +1,11 @@
-import { Exclude } from 'class-transformer';
-import { Column, Entity, PrimaryGeneratedColumn, ManyToOne } from 'typeorm';
-import { TaskStatus } from './task-status.enum';
-import { User } from '../auth/user.entity';
+import { Exclude } from "class-transformer";
+import { Column, Entity, PrimaryGeneratedColumn, ManyToOne } from "typeorm";
+import { TaskStatus } from "./task-status.enum";
+import { User } from "@/auth/schema";
 
 @Entity()
 export class Task {
-  @PrimaryGeneratedColumn('uuid')
+  @PrimaryGeneratedColumn("uuid")
   id: string;
 
   @Column()
@@ -17,7 +17,11 @@ export class Task {
   @Column()
   status: TaskStatus;
 
-  @ManyToOne((_type) => User, (user) => user.tasks, { eager: false })
+  @ManyToOne(
+    _type => User,
+    user => user.tasks,
+    { eager: false },
+  )
   @Exclude({ toPlainOnly: true })
   user: User;
 }
