+++ src/tasks/task.repository.ts
@@ -1,34 +1,34 @@
-import { Repository, DataSource } from 'typeorm';
 import {
   Injectable,
   InternalServerErrorException,
   Logger,
-} from '@nestjs/common';
-import { CreateTaskDto } from './dto/create-task.dto';
-import { GetTaskFilterDto } from './dto/get-tasks.dto';
-import { Task } from './task.entity';
-import { TaskStatus } from './task-status.enum';
-import { User } from '../auth/user.entity';
+} from "@nestjs/common";
+import { DataSource, Repository } from "typeorm";
+import { User } from "@/auth/schema";
+import { CreateTaskDto } from "./dto/create-task.dto";
+import { GetTaskFilterDto } from "./dto/get-tasks.dto";
+import { Task } from "./task.entity";
+import { TaskStatus } from "./task-status.enum";
 
 @Injectable() // here
 export class TaskRepository extends Repository<Task> {
-  private logger = new Logger('TaskRepository');
+  private logger = new Logger("TaskRepository");
   constructor(private dataSource: DataSource) {
     super(Task, dataSource.createEntityManager());
   }
 
   async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
     const { status, search } = filterDto;
-    const query = this.createQueryBuilder('task');
+    const query = this.createQueryBuilder("task");
     query.where({ user });
 
     if (status) {
-      query.andWhere('task.status=:status', { status });
+      query.andWhere("task.status=:status", { status });
     }
 
     if (search) {
       query.andWhere(
-        '(LOWER(task.title) LIKE LOWER(:search) OR LOWER(task.description) LIKE LOWER(:search))',
+        "(LOWER(task.title) LIKE LOWER(:search) OR LOWER(task.description) LIKE LOWER(:search))",
         { search: `%${search}%` },
       );
     }
